+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [vNext]
+- Fixed `SolutionAttribute` to handle empty configuration file
 
 ## [0.12.3] / 2018-11-29
 - Fixed `EnvironmentInfo.Variables` not to be cached